extract sub-modules from json into separate files;
