+++ ergo-lib/src/chain/json/transaction.rs
@@ -1,0 +1,46 @@
+use crate::chain::transaction::{DataInput, Input};
+use crate::chain::{ergo_box::ErgoBox, transaction::TxId};
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, PartialEq, Debug, Clone)]
+pub struct TransactionJson {
+    #[cfg_attr(feature = "json", serde(rename = "id"))]
+    pub tx_id: TxId,
+    /// inputs, that will be spent by this transaction.
+    #[cfg_attr(feature = "json", serde(rename = "inputs"))]
+    pub inputs: Vec<Input>,
+    /// inputs, that are not going to be spent by transaction, but will be reachable from inputs
+    /// scripts. `dataInputs` scripts will not be executed, thus their scripts costs are not
+    /// included in transaction cost and they do not contain spending proofs.
+    #[cfg_attr(feature = "json", serde(rename = "dataInputs"))]
+    pub data_inputs: Vec<DataInput>,
+    #[cfg_attr(feature = "json", serde(rename = "outputs"))]
+    pub outputs: Vec<ErgoBox>,
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::chain::transaction::unsigned::UnsignedTransaction;
+    use crate::chain::transaction::Transaction;
+    use proptest::prelude::*;
+
+    proptest! {
+
+        #[test]
+        fn tx_roundtrip(t in any::<Transaction>()) {
+            let j = serde_json::to_string(&t)?;
+            eprintln!("{}", j);
+            let t_parsed: Transaction = serde_json::from_str(&j)?;
+            prop_assert_eq![t, t_parsed];
+        }
+
+        #[test]
+        fn unsigned_tx_roundtrip(t in any::<UnsignedTransaction>()) {
+            let j = serde_json::to_string(&t)?;
+            eprintln!("{}", j);
+            let t_parsed: UnsignedTransaction = serde_json::from_str(&j)?;
+            prop_assert_eq![t, t_parsed];
+        }
+
+    }
+}
