+++ ergo-lib/src/chain/transaction/input.rs
@@ -28,7 +28,7 @@ pub struct UnsignedInput {
     #[cfg_attr(
         feature = "json",
         serde(rename = "extension",),
-        serde(with = "crate::chain::json::ContextExtensionSerde")
+        serde(with = "crate::chain::json::context_extension::ContextExtensionSerde")
     )]
     pub extension: ContextExtension,
 }
